Fix timestamp type to Long and use correct DATE_TAKEN import

+++ android/src/main/kotlin/com/fluttercandies/photo_manager/core/PhotoManager.kt
@@ -255,7 +255,7 @@ class PhotoManager(private val context: Context) {
         }
     }
 
-    fun updateDateTaken(assetId: String, timestamp: Int, resultHandler: ResultHandler) {
+    fun updateDateTaken(assetId: String, timestamp: Long, resultHandler: ResultHandler) {
         try {
             val result = dbUtils.updateDateTaken(context, assetId, timestamp)
             resultHandler.reply(result)
